+++ apps/rosetta/riflib/rifdock_tasks/HackPackTasks.cc
@@ -137,7 +137,7 @@ HackPackTask::return_search_point_with_rotss(
 
             std::vector<float> scores;
             packed_results[ ipack ].score = rdd.packing_objectives[rif_resl_]->score_with_rotamers( *tscene, scores, packed_results[ ipack ].rotamers() );
-            packed_results[ ipack ].sasa = (uint64_t) ( scores[3] / SASA_SUBVERT_MULTIPLIER );
+            packed_results[ ipack ].sasa = (uint16_t) ( scores[3] / SASA_SUBVERT_MULTIPLIER );
 
 
         } catch( std::exception const & ex ) {
